requirements.txt
Jinja2
Pillow
Flask
flask-wtf
bcrypt
markdown
python-slugify
